Refactor BlockingApplicationRunner
