+++ src/main/kotlin/no/nav/syfo/application/mq/MqUtils.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.util
+package no.nav.syfo.application.mq
 
 import com.ibm.mq.jms.MQConnectionFactory
 import com.ibm.msg.client.jms.JmsConstants
