+++ src/main/resources/db/migration/V4_4__add_timestamps_dialogmeldingopplysninger.sql
